src/requirements.txt
python-dotenv==0.19.0
python-telegram-bot==13.7
pymongo==3.12.0
requests==2.22.0